+++ foundation-runtimes/kotlinx/collections/interoperable/src/commonMain/kotlin/kotlinx/collections/interoperable/Collection.kt
@@ -1,0 +1,5 @@
+package kotlinx.collections.interoperable
+
+import kotlin.collections.Collection as KCollection
+
+expect interface Collection<out E> : Iterable<E>, KCollection<E>