+++ foundation-runtimes/kotlinx/collections/interoperable/src/commonMain/kotlin/kotlinx/collections/interoperable/Iterable.kt
@@ -1,0 +1,5 @@
+package kotlinx.collections.interoperable
+
+import kotlin.collections.Iterable as KIterable
+
+expect interface Iterable<out E> : KIterable<E>