+++ foundation-runtimes/kotlinx/collections/interoperable/src/jsTest/kotlin/JsIteratorTest.kt
@@ -1,0 +1,35 @@
+import ArrayInteroperabilityTest.Company
+import expect.requireFromRootDir
+import kotlinx.collections.interoperable.List
+import kotlinx.collections.interoperable.listOf
+import kotlin.test.Ignore
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class JsIteratorTest {
+//    val me = requireFromRootDir<Me>("kotlinx/collections/interoperable/src/jsTest/resources/me.js")
+//    val companies = listOf("google", "asoft", "yahoo").map { Company(it) }
+//
+//    @Test
+//    @Ignore
+//    fun should_run_a_file_in_resources() {
+//        println("\n\n")
+//        val stuff = me.doStuff(companies.toArray())
+//        println(stuff)
+//    }
+//
+//    @Test
+//    @Ignore
+//    fun can_convert_a_list_to_an_array() {
+//        println("\n\n")
+//        println(me.convertToArray(companies))
+//    }
+//
+//    @Test
+//    @Ignore
+//    fun can_call_unbounded() {
+//        val c = Calc(2)
+//        val res = me.addUnbounded(c, 2)
+//        assertEquals(4, res)
+//    }
+}