+++ foundation-runtimes/terminal/console/src/commonMain/kotlin/terminal/ConsoleTerminalOutput.kt
@@ -4,5 +4,5 @@ class ConsoleTerminalOutput(
     config: TerminalHistoryConfig = TerminalHistoryConfig()
 ) : TerminalOutput {
     override val history = TerminalHistory(config)
-    override fun <T> executePrint(any: T) = kotlin.io.print(any)
+    override fun executePrint(any: Any?) = kotlin.io.println(any)
 }