+++ foundation-runtimes/terminal/console/src/jsMain/kotlin/terminal/npm/promptsync/PromptSync.kt
@@ -2,4 +2,4 @@ package terminal.npm.promptsync
 
 import platform.utils.common.require
 
-internal val prompt = require<() -> ((String) -> String)>("prompt-sync")()
+//internal val prompt = require<() -> ((String) -> String)>("prompt-sync")()